ToggleReplied 0.4 from Thunderbird
This commit contains the unmodified files from version 0.4 of the
ToggleReplied Thunderbird add-on, downloaded from
https://addons.mozilla.org/thunderbird/downloads/latest/togglereplied-10558/addon-10558-latest.xpi,
which is linked to from the add-on's home page,
https://addons.mozilla.org/en-US/thunderbird/addon/togglereplied-10558/.

The maintainer of the add-on has not released a new version since
2013, the web site for the add-on no longer works, and the
maintainer's email address bounces. I am preparing to fork the
maintainer's code and release a new version with compatibility fixes.