Update JavaScript, supported versions for new JS constructs
The "for each (var i in list)" construct has transitioned from
deprecated to unsupported in the latest Thunderbird and SeaMonkey
builds, replaced with "for (var i of list)". This commit updates the
code to reflect that.

That change necessitates two additional changes.

First of all, Postbox is unsupported until they switch to newer JS
engine. Unfortunately, the JavaScript engine in the current version of
Postbox doesn't support the newest JavaScript constructs. I will
re-introduce Postbox support for this add-on when its JS interpreter
has been updated.

Second, the supported versions of Thunderbird and SeaMonkey have been
updated to both desupport older applications that don't support the
new JavaScript constructs, and support newer application versions that
support only the new constructs. The new range of supported versions
is Thunderbird 20 to 57.0 and SeaMonkey 2.17 to 2.54.