Update ID, version, and credits
Change the ID of the add-on to togglereplied@kamens.us.

Add a developer credit for Jonathan Kamens.

Remove redundant creator and contributor credits.

Change "creator" references in the code of the app itself to Jonathan
Kamens, so people using the add-on who have questions about it know
whom to contact.